+++ examples/replay.js
@@ -1,4 +1,4 @@
-const Replay = require('../replay').Replay;
+const Replay = require('../lib/Replay').Replay;
 
 var replay = new Replay(process.argv[2]);
 if (replay) {